% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density_at_grid}
\alias{density_at_grid}
\title{Compute Density at Grid Points (Fast C++ Version)}
\usage{
density_at_grid(
  centered_kernel_mat_at_grid,
  centered_kernel_self_grid,
  lambda_hat,
  weight_hat_vec
)
}
\arguments{
\item{centered_kernel_mat_at_grid}{Matrix of kernel values at grid points.}

\item{centered_kernel_self_grid}{Self-kernel values at grid points.}

\item{lambda_hat}{Scalar lambda parameter.}

\item{weight_hat_vec}{Vector of weights.}
}
\value{
Density values at grid points.
}
\description{
Compute Density at Grid Points (Fast C++ Version)
}
