% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density_at_sampled_x}
\alias{density_at_sampled_x}
\title{Compute Density at Sampled Points (Fast C++ Version)}
\usage{
density_at_sampled_x(
  centered_kernel_mat_at_sampled,
  lambda_hat,
  weight_hat_vec
)
}
\arguments{
\item{centered_kernel_mat_at_sampled}{Matrix of centered kernel values.}

\item{lambda_hat}{Scalar lambda parameter.}

\item{weight_hat_vec}{Vector of weights.}
}
\value{
Density values at sampled points.
}
\description{
Compute Density at Sampled Points (Fast C++ Version)
}
