% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_approx_dens_or_probs.R
\name{get_corresponding_grids_and_their_weights}
\alias{get_corresponding_grids_and_their_weights}
\title{Get Corresponding Grids and Their Weights}
\usage{
get_corresponding_grids_and_their_weights(sample, grid)
}
\arguments{
\item{sample}{A numeric vector representing the sample points.}

\item{grid}{A numeric vector representing the grid.}
}
\value{
A list containing two elements:
\item{grid_weights}{A data frame with the frequency and weight for each grid point.}
\item{corresponding_grid}{A data frame with the details of corresponding grid intervals.}
}
\description{
This function determines the grid intervals corresponding to each sample point
and calculates the weights for each grid point.
}
\examples{
get_corresponding_grids_and_their_weights(c(1, 3, 5), c(0, 2, 4, 6))
}
