% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{get_dens}
\alias{get_dens}
\title{Compute Normalized Densities at Sampled and Grid Points}
\usage{
get_dens(
  centered_kernel_mat_samples,
  centered_kernel_mat_grids,
  centered_kernel_self_grids,
  samples,
  grids,
  base_measure_weights,
  dimension,
  lambda,
  weight_vec
)
}
\arguments{
\item{centered_kernel_mat_samples}{Centered kernel matrix at sampled points.}

\item{centered_kernel_mat_grids}{Centered kernel matrix at grid points.}

\item{centered_kernel_self_grids}{Self-kernel values at grid points.}

\item{samples}{Sampled x values.}

\item{grids}{Grid x values.}

\item{base_measure_weights}{Vector of base measure weights (used for normalizing constant if dim > 1).}

\item{dimension}{Dimensionality of the data (1 or higher).}

\item{lambda}{Scalar lambda parameter.}

\item{weight_vec}{Vector of weights.}
}
\value{
A list with normalized densities at sampled and grid points.
}
\description{
Compute Normalized Densities at Sampled and Grid Points
}
