% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{get_dens_wo_grid}
\alias{get_dens_wo_grid}
\title{Compute Densities for Sampled Points Without a Grid (Fast C++ Version)}
\usage{
get_dens_wo_grid(
  centered_kernel_mat_at_sampled,
  min_x,
  max_x,
  sampled_x,
  lambda_hat,
  weight_hat_vec
)
}
\arguments{
\item{centered_kernel_mat_at_sampled}{Centered kernel matrix at sampled points.}

\item{min_x}{Minimum x value.}

\item{max_x}{Maximum x value.}

\item{sampled_x}{Sampled x values.}

\item{lambda_hat}{Scalar lambda parameter.}

\item{weight_hat_vec}{Vector of weights.}
}
\value{
Normalized densities at sampled points.
}
\description{
Compute Densities for Sampled Points Without a Grid (Fast C++ Version)
}
