% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_approx_dens_or_probs.R
\name{get_middle_points_grid_R}
\alias{get_middle_points_grid_R}
\title{Get Middle Points of Grid}
\usage{
get_middle_points_grid_R(min, samples, max)
}
\arguments{
\item{min}{The minimum value of the grid.}

\item{samples}{A numeric vector representing the sample points within the grid.}

\item{max}{The maximum value of the grid.}
}
\value{
A numeric vector containing the minimum value, the midpoints of the sample, and the maximum value.
}
\description{
This function calculates the midpoints between adjacent elements of a sample
and includes the minimum and maximum grid boundaries.
}
\examples{
get_middle_points_grid(0, c(2, 4, 6), 8)
}
