% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_approx_dens_or_probs.R
\name{get_s_e_mid_points_of_interval}
\alias{get_s_e_mid_points_of_interval}
\title{Get Start and End Midpoints of Interval}
\usage{
get_s_e_mid_points_of_interval(sample_mid_points, grid, s_grid, e_grid)
}
\arguments{
\item{sample_mid_points}{A numeric vector of midpoints between the sample points.}

\item{grid}{A numeric vector representing the grid.}

\item{s_grid}{The start index of the grid interval.}

\item{e_grid}{The end index of the grid interval.}
}
\value{
A numeric vector containing the start and end midpoints of the interval.
}
\description{
This function finds the start and end midpoints of an interval within a grid,
based on the sample midpoints.
}
\examples{
get_s_e_mid_points_of_interval(c(1, 3, 5), c(0, 2, 4, 6), 2, 3)
}
