% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_true_weights.R
\name{get_true_weights}
\alias{get_true_weights}
\title{Compute Approximation of True Weights for Density Estimation}
\usage{
get_true_weights(sample, grid, true_density_at_grid)
}
\arguments{
\item{sample}{A numeric vector of sampled points.}

\item{grid}{A numeric vector representing the grid over which the density is estimated.}

\item{true_density_at_grid}{A numeric vector of true density values at each grid point.}
}
\value{
A numeric vector of estimated weights (\code{w}) corresponding to the sampled points.
}
\description{
This function estimates the true weights (\code{w}) used in a density estimation model.
It computes the centered kernel matrices, evaluates the density function, and
optimizes the weights using least squares error minimization.
}
\examples{
sample <- seq(-3, 3, length.out = 10)
grid <- seq(-3, 3, length.out = 100)
true_density_at_grid <- dnorm(grid)  # Example: Gaussian density
w_estimates <- get_true_weights(sample, grid, true_density_at_grid)
print(w_estimates)

}
