% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{get_weights}
\alias{get_weights}
\title{Estimate Weights Using the Barzilai-Borwein Method (Optimized with Rcpp)}
\usage{
get_weights(
  lambda,
  tau,
  centered_kernel_mat_at_samples,
  samples,
  base_measure_weights,
  dimension,
  print_trace = FALSE
)
}
\arguments{
\item{tau}{A scalar representing the tau parameter.}

\item{centered_kernel_mat_at_samples}{A matrix representing the centered kernel at sampled points.}

\item{samples}{A numeric vector of sampled points.}

\item{base_measure_weights}{A numeric vector of base measures for sample points.}

\item{print_trace}{Logical; if TRUE, prints progress updates.}

\item{lambdas}{A scalar representing the lambda parameter.}

\item{prior_variance_p_vector}{A numeric vector for prior variance probabilities. Default is NULL.}
}
\value{
A numeric vector of estimated weights.
}
\description{
This function estimates the weight vector using the Barzilai-Borwein method,
a numerical approach to solve nonlinear systems of equations. It optimizes
performance by leveraging Rcpp for computation and adaptive sub-sampling for
better initial values in the Barzilai-Borwein method.
}
