% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{l2_ise}
\alias{l2_ise}
\title{Compute the Integrated Squared Error (ISE) for L2 Norm Given the True Densities}
\usage{
l2_ise(true_density_grid, estimated_density_grid)
}
\arguments{
\item{true_density_grid}{Numeric vector. The true density values evaluated at \code{grid} points.}

\item{estimated_density_grid}{Numeric vector. The estimated density values evaluated at \code{grid} points.}
}
\value{
A numeric scalar representing the computed ISE.
}
\description{
This function calculates the Integrated Squared Error (ISE) between the true and estimated density functions using numerical integration.
}
