% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{mmd_grids}
\alias{mmd_grids}
\title{Compute the Maximum Mean Discrepancy (MMD)}
\usage{
mmd_grids(kernel_matrix_at_grids, true_density_grids, estimated_density_grids)
}
\arguments{
\item{kernel_matrix_at_grids}{Numeric matrix The kernel matrix evaluated at the grid points.}

\item{true_density_grids}{Numeric vector. The true density values evaluated at \code{grid} points.}

\item{estimated_density_grids}{Numeric vector. The estimated density values evaluated at \code{grid} points.}
}
\value{
A numeric scalar representing the estimated MMD on grids.
}
\description{
This function calculates the Maximum Mean Discrepancy (MMD)
}
