% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{mmd_samples}
\alias{mmd_samples}
\title{Compute the Maximum Mean Discrepancy (MMD) using samples}
\usage{
mmd_samples(
  kernel_matrix_at_samples,
  true_density_samples,
  estimated_density_samples,
  base_measure_weights
)
}
\arguments{
\item{kernel_matrix_at_samples}{Numeric matrix The kernel matrix evaluated at the sample points.}

\item{true_density_samples}{Numeric vector. The true density values evaluated at \code{sample} points.}

\item{estimated_density_samples}{Numeric vector. The estimated density values evaluated at \code{sample} points.}

\item{base_measure_weights}{Numeric vector. The base measure weights.}
}
\value{
A numeric scalar representing the estimated MMD on samples.
}
\description{
This function calculates the Maximum Mean Discrepancy (MMD) using samples.
}
