% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{rkhs_se}
\alias{rkhs_se}
\title{Compute the RKHS Squared Error (MSE) in an RKHS Given the True Weights}
\usage{
rkhs_se(w_hat_vec, w_vec, kernel_matrix_at_samples)
}
\arguments{
\item{w_hat_vec}{Numeric vector. Estimated weights.}

\item{w_vec}{Numeric vector. True weights.}

\item{kernel_matrix_at_samples}{Numeric matrix. The kernel matrix evaluated at the sample points.}
}
\value{
A numeric scalar representing the computed MSE.
}
\description{
This function calculates the MSE using a Reproducing Kernel Hilbert Space (RKHS) formulation.
}
