% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_mixture.R
\name{rnorm_mixture}
\alias{rnorm_mixture}
\title{Generate Random Samples from a Normal Mixture Model}
\usage{
rnorm_mixture(n, means, sds, mixture_weights)
}
\arguments{
\item{n}{An integer specifying the number of samples to generate.}

\item{means}{A numeric vector of length 4 specifying the means of the normal components.}

\item{sds}{A numeric vector of length 4 specifying the standard deviations of the normal components.}

\item{mixture_weights}{A numeric vector of length 4 specifying the mixture weights for the components.
The weights should sum to 1.}
}
\value{
A numeric vector of length \code{n} containing the generated samples from the normal mixture model.
}
\description{
This function generates random samples from a mixture of normal distributions.
The user specifies the number of samples, the means, standard deviations,
and mixture weights for each component of the mixture.
}
\examples{
# Example usage:
n <- 1000
means <- c(0, 5, 10, 15)
sds <- c(1, 1, 2, 2)
mixture_weights <- c(0.25, 0.25, 0.25, 0.25)
samples <- normal_mixture(n, means, sds, mixture_weights)
hist(samples, breaks = 30, main = "Histogram of Normal Mixture Samples")
}
