% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{unnormalised_density_samples}
\alias{unnormalised_density_samples}
\title{Compute Unnormalized Density at Sampled Points}
\usage{
unnormalised_density_samples(centered_kernel_mat_samples, lambda, weight_vec)
}
\arguments{
\item{centered_kernel_mat_samples}{Matrix of centered kernel values for samples.}

\item{lambda}{Scalar lambda parameter.}

\item{weight_vec}{Vector of weights.}
}
\value{
Unnormalized density values at sampled points.
}
\description{
Compute Unnormalized Density at Sampled Points
}
