% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centered_kernel_matrix.R
\name{centered_kernel_matrix}
\alias{centered_kernel_matrix}
\title{Compute the Centered Kernel Matrix}
\usage{
centered_kernel_matrix(
  first_vec_kernel,
  second_vec_kernel,
  centering_grid,
  hurst_coef
)
}
\arguments{
\item{first_vec_kernel}{Numeric vector.}

\item{second_vec_kernel}{Numeric vector.}

\item{centering_grid}{Numeric vector.}

\item{hurst_coef}{Numeric scalar.}
}
\value{
A numeric matrix.
}
\description{
This function computes the centered kernel matrix efficiently using C++.
}
