% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{get_weights}
\alias{get_weights}
\title{Estimate Weights Using the Barzilai-Borwein Method (Optimized with Rcpp)}
\usage{
get_weights(
  lambda_hat,
  tau_hat,
  centered_kernel_mat_at_sampled,
  sampled_x,
  min_x,
  max_x,
  prior_variance_p_vector = NULL,
  print_trace = FALSE,
  init = FALSE,
  q_with_base = TRUE,
  with_prob_in_var = TRUE,
  normalised_q = TRUE,
  normalised_p = TRUE,
  p_with_base = FALSE
)
}
\arguments{
\item{lambda_hat}{A scalar representing the estimated lambda parameter.}

\item{tau_hat}{A scalar representing the estimated tau parameter.}

\item{centered_kernel_mat_at_sampled}{A matrix representing the centered kernel at sampled points.}

\item{sampled_x}{A numeric vector of sampled points.}

\item{min_x}{The minimum domain value.}

\item{max_x}{The maximum domain value.}

\item{prior_variance_p_vector}{A numeric vector for prior variance probabilities. Default is NULL.}

\item{print_trace}{Logical; if TRUE, prints progress updates.}

\item{init}{Logical; if TRUE, the weights calculation has a sub-sampling method for initialization.}
}
\value{
A numeric vector of estimated weights.
}
\description{
This function estimates the weight vector using the Barzilai-Borwein method,
a numerical approach to solve nonlinear systems of equations. It optimizes
performance by leveraging Rcpp for computation and adaptive sub-sampling for
better initial values in the Barzilai-Borwein method.
}
